+++ src/main/java/com/simsilica/mathd/Vec3i.java
@@ -1,36 +1,36 @@
 /*
  * $Id$
- * 
+ *
  * Copyright (c) 2015, Simsilica, LLC
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions 
+ * modification, are permitted provided that the following conditions
  * are met:
- * 
- * 1. Redistributions of source code must retain the above copyright 
+ *
+ * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- * 
- * 2. Redistributions in binary form must reproduce the above copyright 
- *    notice, this list of conditions and the following disclaimer in 
- *    the documentation and/or other materials provided with the 
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in
+ *    the documentation and/or other materials provided with the
  *    distribution.
- * 
- * 3. Neither the name of the copyright holder nor the names of its 
- *    contributors may be used to endorse or promote products derived 
+ *
+ * 3. Neither the name of the copyright holder nor the names of its
+ *    contributors may be used to endorse or promote products derived
  *    from this software without specific prior written permission.
- * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS 
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT 
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS 
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE 
- * COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, 
- * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES 
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR 
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) 
- * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, 
- * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED 
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+ * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+ * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
+ * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
@@ -59,7 +59,7 @@ public class Vec3i implements Cloneable, java.io.Serializable {
         this.y = y;
         this.z = z;
     }
-    
+
     public Vec3i( Vector3f v ) {
         this((int)v.x, (int)v.y,(int)v.z);
     }
@@ -77,7 +77,7 @@ public class Vec3i implements Cloneable, java.io.Serializable {
     }
 
     @Override
-    public boolean equals( Object o ) {   
+    public boolean equals( Object o ) {
         if( o == this )
             return true;
         if( o == null )
@@ -87,7 +87,7 @@ public class Vec3i implements Cloneable, java.io.Serializable {
         Vec3i other = (Vec3i)o;
         return x == other.x && y == other.y && z == other.z;
     }
-    
+
     @Override
     public int hashCode() {
         int hash = 37;
@@ -177,7 +177,7 @@ public class Vec3i implements Cloneable, java.io.Serializable {
         z -= k;
         return this;
     }
-    
+
     public final Vec3i add( int i, int j, int k ) {
         return new Vec3i( x + i, y + j, z + k );
     }
@@ -185,30 +185,30 @@ public class Vec3i implements Cloneable, java.io.Serializable {
     public final Vec3i add( Vec3i v ) {
         return new Vec3i( x + v.x, y + v.y, z + v.z );
     }
- 
+
     public final Vec3i subtract( Vec3i v ) {
         return new Vec3i( x - v.x, y - v.y, z - v.z );
-    } 
-    
+    }
+
     public final Vec3i subtract( int vx, int vy, int vz ) {
         return new Vec3i( x - vx, y - vy, z - vz );
-    } 
-    
+    }
+
     public final Vec3i mult( int scale ) {
         return new Vec3i( x * scale, y * scale, z * scale );
     }
-    
+
     public final int getDistanceSq( Vec3i v ) {
         int xd = v.x - x;
         int yd = v.y - y;
         int zd = v.z - z;
-        
+
         return xd * xd + yd * yd + zd * zd;
     }
-      
+
     public final double getDistance( Vec3i v ) {
         return Math.sqrt(getDistanceSq(v));
-    }  
+    }
 
     public Vec3i minLocal( int i, int j, int k ) {
         x = Math.min(x, i);
@@ -216,7 +216,7 @@ public class Vec3i implements Cloneable, java.io.Serializable {
         z = Math.min(z, k);
         return this;
     }
-    
+
     public Vec3i maxLocal( int i, int j, int k ) {
         x = Math.max(x, i);
         y = Math.max(y, j);
@@ -230,7 +230,7 @@ public class Vec3i implements Cloneable, java.io.Serializable {
         z = Math.min(v.z, z);
         return this;
     }
-    
+
     public Vec3i maxLocal( Vec3i v ) {
         x = Math.max(v.x, x);
         y = Math.max(v.y, y);